+++ RELEASE_NOTES.md
@@ -2,6 +2,10 @@
 
 ### RELEASE 0.1.9-alpha
 
+ENHANCEMENTS:
+
+1. Add `enc2float` helper method to convert SPARTN encoded floating point values to floats.
+
 CHANGES:
 
 1. Add temporary override of decode flag for message types that cannot yet be properly decoded (e.g. OCB)