+++ RELEASE_NOTES.md
@@ -9,6 +9,7 @@ ENHANCEMENTS:
 CHANGES:
 
 1. Add temporary override of decode flag for message types that cannot yet be properly decoded (e.g. OCB)
+1. Add `gad_plot.py` example to illustrate how to extract geographic area definitions from SPARTN-1X-GAD messages.
 
 ### RELEASE 0.1.8-alpha